Move enable_pin_uv back to ctap/mod.rs
