+++ src/test_helpers/mod.rs
@@ -13,15 +13,13 @@
 // limitations under the License.
 
 use crate::clock::CtapInstant;
-use crate::ctap::client_pin::{ClientPin, PIN_TOKEN_LENGTH};
 use crate::ctap::command::{
     AuthenticatorAttestationMaterial, AuthenticatorConfigParameters,
     AuthenticatorVendorConfigureParameters, Command,
 };
-use crate::ctap::data_formats::{ConfigSubCommand, PinUvAuthProtocol};
-use crate::ctap::pin_protocol::authenticate_pin_uv_auth_token;
+use crate::ctap::data_formats::ConfigSubCommand;
 use crate::ctap::status_code::Ctap2StatusCode;
-use crate::ctap::{key_material, storage, Channel, CtapState};
+use crate::ctap::{key_material, Channel, CtapState};
 use crate::env::Env;
 
 // In tests where we define a dummy user-presence check that immediately returns, the channel
@@ -61,30 +59,4 @@ pub fn enable_enterprise_attestation(
     state.process_parsed_command(env, config_command, DUMMY_CHANNEL, CtapInstant::new(0))?;
 
     Ok(attestation_material)
-}
-
-pub fn enable_pin_uv(
-    state: &mut CtapState,
-    env: &mut impl Env,
-    pin_uv_auth_protocol: PinUvAuthProtocol,
-    slot_id: usize,
-    client_data_hash: &[u8],
-) -> Result<Vec<u8>, Ctap2StatusCode> {
-    let key_agreement_key = crypto::ecdh::SecKey::gensk(env.rng());
-    let pin_uv_auth_token = [0x91; PIN_TOKEN_LENGTH];
-    let client_pin = ClientPin::new_test(
-        env,
-        slot_id,
-        key_agreement_key,
-        pin_uv_auth_token,
-        pin_uv_auth_protocol,
-    );
-    state.set_client_pin_for_test(client_pin);
-    storage::set_pin(env, slot_id, &[0x88; 16], 4)?;
-
-    Ok(authenticate_pin_uv_auth_token(
-        &pin_uv_auth_token,
-        client_data_hash,
-        pin_uv_auth_protocol,
-    ))
 }
